added Makefile to simplify setup
The Makefile basically guides the user through the entire build process
for the container by a simple `make image` and starts the entire circus
:circus_tent: with `make notebook`. I know this is a two year old
repository, but after I just watched your
[talk at Con2016](https://www.youtube.com/watch?v=7_DX1EGKZXY), I just
had to walk through the examples for the heck of it. :grin:

The `make notebook` spawns the notebook under a non-root user, which is
why I chmod-ed `/home` in my previous commit and create a `/home/you`
directory which I set to `$HOME` in order to start an instance of the
container as a user with the same uid and gid as the user on the host.
On Linux this at least, produces files on mounted volumes that have the
uid and gid of the host user. :smile: